Sửa định dạng thư mục
Định dạng BT1
