+++ PhanDoanCuong/BT1/README.txt
@@ -1,0 +1,7 @@
++Bài tập nhỏ tìm số nguyên tố lớn nhất nhỏ hơn số N.
+ +Input: Số N
+ +Output: X là số nguyên tố lớn nhất nhỏ hơn số N.
+ +Cài đặt:
+ ++Hàm kiểm tra một số có phải là số nguyên tố không.
+ ++Hàm trả về số nguyên tố ứng với giá trị đầu vào.
+ ++Chương trình sử dụng JUnitTest.
